requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
frozenlist==1.3.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
multidict==6.0.4
openai==0.26.5
Pillow==9.4.0
python-telegram-bot==20.1
requests==2.28.2
rfc3986==1.5.0
sniffio==1.3.0
tqdm==4.64.1
urllib3==1.26.14
yarl==1.8.2
